+++ src/tree/tree.h
@@ -16,11 +16,11 @@
 #include "../common/colors.h"
 #include "operations.h"
 
-const int MAX_N_NODES = 50000;  // max number of nodes in AST
-const int MAX_TREE    = 500000; // max len of a string-written tree in a file
-const int MAX_OP      = 200;    // max len of an operator
+const size_t MAX_N_NODES = 50000;  // max number of nodes in AST
+const size_t MAX_TREE    = 500000; // max len of a string-written tree in a file
+const size_t MAX_OP      = 200;    // max len of an operator
 
-const int NAMETABLE_CAPACITY = 15;
+const size_t NAMETABLE_CAPACITY = 15;
 
 const double EXPONENT = 2.718281828;
 const double PI       = 3.141592654;
